Refactor code and move files into component dir.
